Move runtime metadata and stack-value declarations to guest staging
Move guest-owned analysis and runtime metadata annotations together with the
stack-value intrinsic declarations out of core and into guest staging. Builder
code still interprets the metadata and registers the intrinsic plugins, but it
does not own the declarations used by guest runtime code.

Hosted annotation consumers temporarily retain their existing access paths
until generated annotation-value wrappers are available. Graph-builder
plugins, compiler nodes, and lowering remain builder-owned.
