+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/core/heap/RestrictHeapAccess.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core.heap;
+package com.oracle.svm.guest.staging.core.heap;
 
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
