+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/core/heap/UnknownPrimitiveField.java
@@ -22,7 +22,7 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package com.oracle.svm.core.heap;
+package com.oracle.svm.guest.staging.core.heap;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
